HV:Treewide:Update the type of vcpu id as uint16_t
In the hypervisor, virtual cpu id is defined as "int" or "uint32_t"
type in the hypervisor. So there are some sign conversion issues
about virtual cpu id (vcpu_id) reported by static analysis tool.
Sign conversion violates the rules of MISRA C:2012.

BTW, virtual cpu id has different names (vcpu_id, cpu_id, logical_id)
 for different modules of HV, its type is defined as "int" or "uint32_t"
in the HV. cpu_id type and logical_id type clean up will be done in
other patchs.

V1-->V2:
         More clean up the type of vcpu id;
         "%hu" is for vcpu id in the print function.

+++ hypervisor/arch/x86/ept.c
@@ -275,7 +275,7 @@ void unregister_mmio_emulation_handler(struct vm *vm, uint64_t start,
 int dm_emulate_mmio_post(struct vcpu *vcpu)
 {
 	int ret = 0;
-	int cur = vcpu->vcpu_id;
+	uint16_t cur = vcpu->vcpu_id;
 	union vhm_request_buffer *req_buf;
 
 	req_buf = (union vhm_request_buffer *)(vcpu->vm->sw.io_shared_page);
